changed stat values from int to double

+++ Troops/BaseInterfaces/IAccuracy.cs
@@ -8,6 +8,6 @@ namespace WarSimulator.Troops.BaseInterfaces
 {
     public interface IAccuracy
     {
-        public int Accuracy { get; set; }
+        public double Accuracy { get; set; }
     }
 }
