+++ Troops/BaseInterfaces/ITroop.cs
@@ -8,10 +8,10 @@ namespace WarSimulator.Troops.BaseInterfaces
 {
     public interface ITroop
     {
-        public int Attack { get; set; }
-        public int Defence { get; set; }
-        public int Speed { get; set; }
-        public int Life { get; set; }
-        public int Price { get; set; }
+        public double Attack { get; set; }
+        public double Defence { get; set; }
+        public double Speed { get; set; }
+        public double Life { get; set; }
+        public double Price { get; set; }
     }
 }
